:recycle: merge files
