--- internal/domain/lifeevent.go
@@ -1,48 +1,0 @@
-package domain
-
-import (
-	"time"
-
-	"github.com/google/uuid"
-	"github.com/hackathon-22-winter-01/hackathon-22-winter-01_backend/pkg/consts"
-)
-
-// LifeEvent プレイヤーのライフに関する情報
-type LifeEvent struct {
-	ID        uuid.UUID
-	Type      LifeEventType
-	Diff      float32
-	CreatedAt time.Time
-}
-
-// LifeEventType ライフイベントの種類
-type LifeEventType uint8
-
-const (
-	LifeEventTypeDamaged LifeEventType = iota
-	LifeEventTypeHealed
-)
-
-func NewLifeEvent(id uuid.UUID, typ LifeEventType, diff float32, createdAt time.Time) *LifeEvent {
-	return &LifeEvent{
-		ID:        id,
-		Type:      typ,
-		Diff:      diff,
-		CreatedAt: createdAt,
-	}
-}
-
-func CalculateLife(events []*LifeEvent) float32 {
-	life := consts.MaxLife
-
-	for _, e := range events {
-		switch e.Type {
-		case LifeEventTypeDamaged:
-			life -= e.Diff
-		case LifeEventTypeHealed:
-			life += e.Diff
-		}
-	}
-
-	return life
-}
